requirements.txt
django==4.2.19
djangorestframework==3.15.2
django-cors-headers==4.7.0